requirements.txt
blinker
click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
flask-cors  
psycopg2-binary  
pytest
Flask-SQLAlchemy
python-dotenv